src/service_1/requirements.txt
# =========================
# Mains
# =====
django==5.1.6
django-ninja==1.3.0
gunicorn==23.0.0

requests==2.32.3

grpcio==1.71.0
grpcio-tools==1.71.0
mypy-protobuf==3.6.0

kafka-python==2.1.2

opentelemetry-sdk==1.30.0
opentelemetry-exporter-otlp==1.30.0
opentelemetry-instrumentation-logging==0.51b0
opentelemetry-instrumentation-django==0.51b0
opentelemetry-instrumentation-grpc==0.51b0
opentelemetry-instrumentation-requests==0.51b0
opentelemetry-instrumentation-sqlite3==0.51b0
opentelemetry-instrumentation-kafka-python==0.51b0

# =========================
# Devs
# =====
pre-commit==3.5.0
ruff==0.9.9
types-requests==2.32.0.20250306
